refactor(core,loader,plugin): tighten Spanned<Posting> per review
Addresses all eight findings from the deep PR review of #1151. No new
feature; just sharpening the rough edges in the foundation before
downstream PRs land on it.

Must-fix

- Bump cache header from v4 to v5. The on-disk rkyv encoding of
  Transaction.postings changed shape; without the bump, stale caches
  would deserialize as junk. The header check short-circuits and
  forces a rebuild.
- Loader propagation now hard-errors on file_id > u16::MAX instead of
  silently mapping the 65,536th file onto SYNTHESIZED_FILE_ID. Matches
  the existing debug_assert in Spanned::with_file_id.
- Sweep all literal Span::new(0, 0) callsites to Span::ZERO so the
  named constant is the single canonical way to spell the synthesis
  sentinel.

Spanned<T> semantics

- PartialEq / Eq / Hash on Spanned<T> now delegate to the inner value
  only. Two values with the same content but different source
  locations are equal. Consumers that need location-sensitive
  equality compare .span / .file_id explicitly. Pinned with
  test_spanned_eq_ignores_location plus HashSet round-trip coverage.

Wire format

- SourceSpan offsets widened from u32 to u64. The previous u32
  silently clamped at u32::MAX for files >4 GiB; u64 is
  target-portable and defensive at trivial cost.
- data_to_spanned_posting now surfaces an explicit SpanOverflow
  ConversionError on the (hypothetical) 32-bit-host overflow path
  instead of casting u64 to usize blindly.

API

- Renamed Transaction::with_posting(Posting) to
  with_synthesized_posting(Posting) and added a new
  with_posting(Spanned<Posting>) that pushes a real wrapper directly.
  Silent synthesis is now visible at every call site. All ~800
  in-tree call sites updated.
- Posting::with_cost / with_price / with_flag rustdoc cross-link
  Spanned::map for the wrapper-preserving pattern.

Plugin trust

- apply_plugin_ops now calls sanitize_inner_posting_spans on the
  inner postings of every Modify / Insert directive. Any inner span
  whose file_id doesn't resolve in SourceMap, or whose byte range
  exceeds the file's length, collapses to Spanned::synthesized. A
  misbehaving plugin can no longer smuggle out-of-bounds spans into
  the LSP edit path.

Docs

- rustledger-wasm: TS interface comment notes the deliberate
  flattening of Spanned<Posting> to plain Posting at the JS boundary.

Verified: cargo check --workspace --all-features --all-targets,
cargo clippy --workspace --all-features --all-targets -- -D warnings,
cargo test --workspace --all-features (56 suites, 0 failures),
cargo fmt --all -- --check.

Refs #1142

+++ crates/rustledger-plugin-types/src/lib.rs
@@ -444,16 +444,17 @@ pub struct TransactionData {
 /// brand-new posting, leave [`PostingData::span`] as `None` and the host
 /// will mark it `SYNTHESIZED_FILE_ID`.
 ///
-/// Byte offsets are stored as `u32` so the wire format is stable across
-/// 32-bit (WASM) and 64-bit (host) targets — beancount files >4 GiB do
-/// not exist in practice. The contents are otherwise opaque to plugin
-/// code: do not synthesize spans by guessing offsets.
+/// Byte offsets are stored as `u64` so the wire format is stable
+/// across 32-bit (WASM) and 64-bit (host) targets, and so very large
+/// concatenated source trees (includes-of-includes) cannot silently
+/// overflow. The contents are otherwise opaque to plugin code: do
+/// not synthesize spans by guessing offsets.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
 pub struct SourceSpan {
     /// Start byte offset within the file (inclusive).
-    pub start: u32,
+    pub start: u64,
     /// End byte offset within the file (exclusive).
-    pub end: u32,
+    pub end: u64,
     /// Source file index in the host's source map.
     pub file_id: u16,
 }
